+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/keyvalue/impl/NamespaceMappingKeyValueService.java
@@ -214,7 +214,7 @@ public class NamespaceMappingKeyValueService extends ForwardingObject implements
     }
 
     @Override
-    public void addGarbageCollectionSentinelValues(TableReference tableRef, Set<Cell> cells) {
+    public void addGarbageCollectionSentinelValues(TableReference tableRef, Iterable<Cell> cells) {
         delegate().addGarbageCollectionSentinelValues(tableRef, cells);
     }